refactor(test_functions): migrate Cheng2D to YAML spec
Convert the Cheng2D function from the legacy ABC-based system to a
YAML-driven spec as part of the Phase 2 fixed-dimension migration.

- Add cheng2d.yaml with Cheng2010 input spec
- Strip ABC boilerplate from `cheng2d.py`, keeping only `evaluate()`
- Fix year typo in opening paragraph (2002 to 2010)
- Revise title and opening paragraph to describe the function type
- Fix matplotlib labels to use raw strings for correct LaTeX rendering

Ref: #503
